Move tabel to table for better Java mapings
