+++ src/main/java/org/leycm/table/Table.java
@@ -1,4 +1,4 @@
-package org.leycm.tabel;
+package org.leycm.table;
 
 import java.util.List;
 import java.util.function.Predicate;
